+++ frontend/app/player/web/managers/DOM/SelectionManager.ts
@@ -9,15 +9,12 @@ export default class SelectionManager extends ListWalker<SelectionChange> {
   }
 
   private selected: [{ id: number, node: Element } | null, { id: number, node: Element } | null] = [null, null];
-  private markers: Element[] = []
 
-  clearSelection() {
-    this.selected[0] && this.screen.overlay.removeChild(this.selected[0].node) && this.selected[0].node.remove();
-    this.selected[1] && this.screen.overlay.removeChild(this.selected[1].node) && this.selected[1].node.remove();
-    this.markers.forEach(marker => marker.remove())
+  public clearSelection = () => {
+    if (this.selected[0] === null && this.selected[1] === null) return;
 
+    this.screen.clearSelection()
     this.selected = [null, null];
-    this.markers = [];
   }
 
   move(t: number) {
@@ -25,6 +22,7 @@ export default class SelectionManager extends ListWalker<SelectionChange> {
     if (!msg) {
       return;
     }
+
     // in theory: empty selection or selection removed
     if (msg.selectionStart <= 0) {
       this.clearSelection()
@@ -48,26 +46,22 @@ export default class SelectionManager extends ListWalker<SelectionChange> {
 
       Object.assign(endPointer.style, {
         top: endCoords.top + 'px',
-        left: (endCoords.left + (endCoords.width / 2) + 3) + 'px',
-        width: (endCoords.width / 2) + 'px',
+        right: (endCoords.right) + 'px',
+        width: (endCoords.width) + 'px',
         height: endCoords.height + 'px',
-        borderRight: '2px solid blue',
         position: 'absolute',
         boxShadow: '1px 4px 1px -2px blue',
       });
       Object.assign(startPointer.style, {
         top: startCoords.top + 'px',
-        left: (startCoords.left - 3) + 'px',
-        width: (startCoords.width / 2 ) + 'px',
+        left: (startCoords.left) + 'px',
+        width: (startCoords.width) + 'px',
         height: startCoords.height + 'px',
-        borderLeft: '2px solid blue',
         position: 'absolute',
         boxShadow: '1px 4px 1px -2px blue',
       });
 
-      this.markers.push(startPointer, endPointer);
-      this.screen.overlay.appendChild(startPointer);
-      this.screen.overlay.appendChild(endPointer);
+      this.screen.createSelection(startPointer, endPointer);
 
       this.selected = [{ id: msg.selectionStart, node: startPointer }, { id: msg.selectionEnd, node: endPointer }];
     }