+++ tracker/tracker/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 7.0.1
+
+- fix time inputs capturing
+- add option to disable network tracking inside iframes
+
 # 7.0.0
 
 - **[breaking]** added gzip compression to large messages